depend on Discord from master branch of nixpkgs
When discord updates their client to a new version, the auto-update
mechanism disallows using the prior version. This works reasonably
when you buy into the auto-update but it's a no go on NixOS. The
version bump for discord was just merged a while ago onto master
of nixpkgs but it hasn't beeen released to the channels yet because
of hydra build failures unrelated to the upgrade.

So this commit adds an input for nixpkgs master branch and overlays
discord into `pkgs`.