use helix flake instead of overlay
This is the best of both worlds now with
https://github.com/helix-editor/helix/pull/1659: I can use a custom
set of grammars in a dirty branch on my fork and pull it in without
writing a custom nix expression to package it. The prior solution
here with the overlay (now deleted) was a bit slow to fetch all of
the grammars. `grammars.nix` from the 1659 PR seems to be faster
anecdotally.