switch from docker to podman
The main draw to podman for me is that it doesn't use a daemon by
default, and most importantly: it doesn't use containerd at all.
`containerd-shim` doesn't die when it receives SIGTERM, meaning
that the OS hangs when doing a shutdown or reboot while waiting for
containerd-shim to respond to the graceful kill request. The OS
eventually brutal-kills the process and it just waits there for
however long which is very annoying. With podman installed, my
reboot is comparatively instant.

Docker-compose appears to still work on podman after the podman
v3.0.0 release, so I'm not sacrificing much.