remove erlang-ls from overall packages
I'm hacking on my fork on this branch

    https://github.com/the-mikedavis/erlang_ls/tree/632d9823c655e369c90695b8c4556f018d9e8abd

which has a commit that sets up a nix flake, so I can just

    nix profile install path/to/that/dir

and `nix profile remove <index of package in profile>` when removing.
This ends up being a faster workflow when hacking on a package, and
it's not too bad of a Nix sin because you can still install the
package manually on another machine as long as the repository is
pushed. Plus eventually I can just depend on that flake in my dotfiles
flake if I want.