spring cleaning for ~/: use home-manager impermanence
nix-community/impermanence is the input that lets me specify my root
file system as tmpfs and bind-mount persistent data out of a disk.
It's nice to do this because it forces you to be rigorous and explicit
about what data you let persist between reboots. Without it you can be
tempted to make quick hacks, and quick hacks are bad because they
can't be easily replicated across machines.

This change mostly focuses on using the impermanance project's
home-manager integration, which allows you to persist files from
a user's directory. So I have a commented list of directories and
files that persist across reboots and the rest are wiped.

Also included is a change that marks all directories as safe for
root's `git`. This was annoying, I'd have to

    sudo git config --global --add safe.directory '*'
    sudo nixos-rebuild switch --flake .

any time I wanted to change dotfiles config after any reboot. With
this change (that lives in ./users/root.nix), I only have to run
the second command there.