git: Setup SSH commit/tag signing
GitHub now supports SSH-signed commits[^1]. This guide[^2] came out
around the time that git itself supported using SSH to sign commits.
We can use the tricks there to set up SSH signing and verify commits
locally before trying to push them to a remote.

Ideally in the long run I should not have to hard-code the public
key ID like this. Maybe it's fine as long as I take the key ID as
an argument and have a constant per-machine.

[^1]: https://github.blog/changelog/2022-08-23-ssh-commit-verification-now-supported/
[^2]: https://calebhearth.com/sign-git-with-ssh