Add distrobox
Bazel doesn't like running on NixOS. It's good about build inputs
and outputs but not good about package inputs: it isn't rigorous about
the dependencies you need to install on a system in order to perform
a build. On Linux it tries to execute downloaded ELFs which doesn't
work in NixOS.

Even if you can build successfully locally, Remote Build Execution
(RBE) doesn't work because the remote runner tries to execute bash
out of the Nix store for example which doesn't exist on a non-NixOS
machine (and wouldn't necessarily exist on a remote NixOS machine
anyways).

Now I use distrobox which lets me have a kind of "sidecar OS" that
integrates very well with the host OS. I say `distrobox enter` and
suddenly that shell is in ArchLinux with all the correct directories
mounted. It's almost as seemless as the Nix shell.

For work:

    distrobox create --name rabbit --image archlinux:latest --additional-flags "--pids-limit 65536"

The pids limit is important so that Bazel can spawn enough threads in
a build. It sees that I have many CPU cores and tries to estimate a
good number of builds to run concurrently but the cgroup nprocs limit
(which can be found with `cat /sys/fs/cgroup/pids.max`) limits it to
1024 or 2048 or something else low by default, causing errorno 11 for
`fork` and failing builds. Bumping that number up to something high but
lower or around `ulimit -u` fixes the issue.

So far distrobox is quite nice: it has the permanence of a virtual
machine but the seemless host<->container interaction of a process
running directly on the machine (which it is).