mango2: Upgrade GPU
Switched from a Radeon RX580 to a RX7900-XT. This didn't work out
of the box: the new card is newer than the latest NixOS release. Some
recent upstream patches to MESA and other graphics related
programs/libs and the linux kernel are needed to run. Running from
unstable and switching to the latest linuxPackages_testing (kernel
6.2.0 vs 5.15.x) now works though.

I'm not sure which of these `"amdgpu"` kernel module additions is
necessary. The RX580 was on the old GCN architecture which worked
out of the box.