Impermanence: Try symlink method for Documents & src/
With a simple test that creates a 10GB file:

    dd if=/dev/zero of=example.bin bs=1G count=10

I see ~1,000 MB/s on the disk when under /nix/persist/ and around
500 MB/s on any bind mounted directory. Using symlinks yields the same
speed as being on the disk itself.

This will hopefully also solve issues I see where running large numbers
of unit tests with the Rust compiler or creating a project with the
Gleam compiler errors weirdly. I assume that was caused some odd
interaction between fuse and file system accesses.