Restore official spotify native app
Previously I replaced this with calls to

     nohup chromium --app=https://open.spotify.com --enable-features=UseOzonePlatform --ozone-platform=wayland &; disown %1

because XWayland windows would not register mouse motions on my second
monitor (which is vertical) in places higher than my first monitor
(which is horizontal). That's exactly where I want to keep my Spotify
window, so I had to replace the XWayland official native app with the
chromium wrapper running on ozone.

It turns out that the odd mouse behavior was caused because my second
monitor was positioned with negative coordinates. 041c68d switched
the coordinates of the monitor so that the second monitor was at 0,0
and the first monitor shifted over to 2160,875 which fixed the mouse
behavior.

Now we can switch back to the official native app.