darwin: Update flake and inputs
I was on ancient nixpkgs and home-manager, this updates to the latest
expected format. Error messages pointed me to:
<https://nix-community.github.io/home-manager/release-notes.html#sec-release-22.11-highlights>
for how to update.

Impetus was to upgrade lazygit to the point where it magically knows how
to get Helix to jump to an exact line from the diff viewer. (..mission
accomplished..)